x68k: shrink size of ite_decgraph2ascii[] table.

+++ sys/arch/x68k/dev/ite.c
@@ -2319,7 +2319,7 @@ itecheckwrap(struct ite_softc *ip)
  * A convertion table from DEC special graphics characters to ASCII characters.
  * Mostly for box drawing on sysinst(8).
  */
-const int ite_decgraph2ascii[128] = {
+const uint8_t ite_decgraph2ascii[128] = {
 	/* same as ASCII from 0x00 to 0x5e */
 	0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07,
 	0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f,
